+++ app/src/main/java/com/polyarniq/calkgp/ui/notary/NotaryFragment.kt
@@ -67,7 +67,7 @@ class NotaryFragment : Fragment() {
             val result = DutyCalculator.calculateNotary(option.type, amount)
             findNavController().navigate(
                 R.id.action_notary_to_result,
-                bundleOf("amount" to result.amount, "legalRef" to result.legalReference, "description" to result.description)
+                bundleOf("amount" to result.amount.toFloat(), "legalRef" to result.legalReference, "description" to result.description)
             )
         }
     }
